windows: make installed-driver detection work even if device has never been plugged in
Windows SetupAPI lists devices known to Windows, but a driver can be bound to a device
without that device actually being in Windows' device database, if you install a driver
for it and haven't plugged the device in yet.

It turns out we can check `HKLM:\SYSTEM\DriverDatabase\DeviceIds\USB\VID_1D50&PID_6018&MI_04`
and `HKLM:\SYSTEM\DriverDatabase\DeviceIds\USB\VID_1D50&PID_6017`, which each list INF names
for drivers bound to devices with those hardware IDs. So that's what we do now.